Modularization of EDA page --- DONE
